Criei a interface AçoesPersonagem, que tem os metodos em comum entre as interfaces AcoesHerois e AcoesViloes
