--- src/personagens/herois/AcoesPersonagem.java
@@ -1,19 +1,0 @@
-package personagens.herois;
-
-import itens.Item;
-import personagens.viloes.Vilao;
-
-public interface AcoesPersonagem {
-    void contarHistoria();
-    void poderEspecial1(Vilao vilao);
-    void poderEspecial2();
-    void ataquePadrao(Vilao vilao);
-    void aumentarMana(int quantidadeMana);
-    void diminuirMana(int quantidadeMana);
-
-    void aumentarVida(int quantidadeVida);
-    void diminiurVida(int quantidadeVida);
-    void aumentarVidaRodada();
-    void aumentarManaRodada();
-    void adicionarItem(Item item);
-}
