+++ src/personagens/viloes/AcoesViloes.java
@@ -1,0 +1,7 @@
+package personagens.viloes;
+
+import personagens.herois.Heroi;
+
+public interface AcoesViloes {
+    void tomadadeDecisao();
+}
