--- src/personagens/viloes/Obstaculo.java
@@ -1,8 +1,0 @@
-package personagens.viloes;
-
-import personagens.herois.Heroi;
-
-public interface Obstaculo  {
-    void ataquePadrao(Heroi heroi);
-    void diminiurVida(int quantidadeVida);
-}
